+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - 👷(docker) add arm64 platform support for image builds
 - ✨(global) add create file from template feature
 - ✨(global) add FRONTEND_CSS_URL and FRONTEND_JS_URL settings
+- ✨(backend) add a download action returning the media url
 
 ### Fixed
 
@@ -328,5 +329,3 @@ and this project adheres to
 [v0.2.0]: https://github.com/suitenumerique/drive/releases/v0.2.0
 [v0.1.1]: https://github.com/suitenumerique/drive/releases/v0.1.1
 [v0.1.0]: https://github.com/suitenumerique/drive/releases/v0.1.0
-
-## [v0.11.1] - 2026-01-13